+++ pkg/kubelet/allocation/allocation_manager_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package allocation
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -25,15 +24,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/component-base/metrics/testutil"
 	"k8s.io/kubernetes/pkg/kubelet/allocation/state"
-	"k8s.io/kubernetes/pkg/kubelet/cm/topologymanager"
-	"k8s.io/kubernetes/pkg/kubelet/eviction"
-	"k8s.io/kubernetes/pkg/kubelet/lifecycle"
-	"k8s.io/kubernetes/pkg/kubelet/metrics"
-	"k8s.io/kubernetes/pkg/kubelet/nodeshutdown"
-	"k8s.io/kubernetes/pkg/kubelet/sysctl"
-	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/tainttoleration"
 )
 
 func TestUpdatePodFromAllocation(t *testing.T) {
@@ -174,212 +165,6 @@ func TestUpdatePodFromAllocation(t *testing.T) {
 			} else {
 				assert.False(t, updated, "updated")
 				assert.Same(t, pod, allocatedPod)
-			}
-		})
-	}
-}
-
-func TestRecordAdmissionRejection(t *testing.T) {
-	metrics.Register()
-
-	testCases := []struct {
-		name   string
-		reason string
-		wants  string
-	}{
-		{
-			name:   "AppArmor",
-			reason: lifecycle.AppArmorNotAdmittedReason,
-			wants: `
-				# HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-				# TYPE kubelet_admission_rejections_total counter
-				kubelet_admission_rejections_total{reason="AppArmor"} 1
-			`,
-		},
-		{
-			name:   "PodOSSelectorNodeLabelDoesNotMatch",
-			reason: lifecycle.PodOSSelectorNodeLabelDoesNotMatch,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="PodOSSelectorNodeLabelDoesNotMatch"} 1
-            `,
-		},
-		{
-			name:   "PodOSNotSupported",
-			reason: lifecycle.PodOSNotSupported,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="PodOSNotSupported"} 1
-            `,
-		},
-		{
-			name:   "InvalidNodeInfo",
-			reason: lifecycle.InvalidNodeInfo,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="InvalidNodeInfo"} 1
-            `,
-		},
-		{
-			name:   "InitContainerRestartPolicyForbidden",
-			reason: lifecycle.InitContainerRestartPolicyForbidden,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="InitContainerRestartPolicyForbidden"} 1
-            `,
-		},
-		{
-			name:   "SupplementalGroupsPolicyNotSupported",
-			reason: lifecycle.SupplementalGroupsPolicyNotSupported,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="SupplementalGroupsPolicyNotSupported"} 1
-            `,
-		},
-		{
-			name:   "UnexpectedAdmissionError",
-			reason: lifecycle.UnexpectedAdmissionError,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="UnexpectedAdmissionError"} 1
-            `,
-		},
-		{
-			name:   "UnknownReason",
-			reason: lifecycle.UnknownReason,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="UnknownReason"} 1
-            `,
-		},
-		{
-			name:   "UnexpectedPredicateFailureType",
-			reason: lifecycle.UnexpectedPredicateFailureType,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="UnexpectedPredicateFailureType"} 1
-            `,
-		},
-		{
-			name:   "node(s) had taints that the pod didn't tolerate",
-			reason: tainttoleration.ErrReasonNotMatch,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="node(s) had taints that the pod didn't tolerate"} 1
-            `,
-		},
-		{
-			name:   "Evicted",
-			reason: eviction.Reason,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="Evicted"} 1
-            `,
-		},
-		{
-			name:   "SysctlForbidden",
-			reason: sysctl.ForbiddenReason,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="SysctlForbidden"} 1
-            `,
-		},
-		{
-			name:   "TopologyAffinityError",
-			reason: topologymanager.ErrorTopologyAffinity,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="TopologyAffinityError"} 1
-            `,
-		},
-		{
-			name:   "NodeShutdown",
-			reason: nodeshutdown.NodeShutdownNotAdmittedReason,
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="NodeShutdown"} 1
-            `,
-		},
-		{
-			name:   "OutOfcpu",
-			reason: "OutOfcpu",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="OutOfcpu"} 1
-            `,
-		},
-		{
-			name:   "OutOfmemory",
-			reason: "OutOfmemory",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="OutOfmemory"} 1
-            `,
-		},
-		{
-			name:   "OutOfephemeral-storage",
-			reason: "OutOfephemeral-storage",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="OutOfephemeral-storage"} 1
-            `,
-		},
-		{
-			name:   "OutOfpods",
-			reason: "OutOfpods",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="OutOfpods"} 1
-            `,
-		},
-		{
-			name:   "OutOfgpu",
-			reason: "OutOfgpu",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="OutOfExtendedResources"} 1
-            `,
-		},
-		{
-			name:   "OtherReason",
-			reason: "OtherReason",
-			wants: `
-                # HELP kubelet_admission_rejections_total [ALPHA] Cumulative number pod admission rejections by the Kubelet.
-                # TYPE kubelet_admission_rejections_total counter
-                kubelet_admission_rejections_total{reason="Other"} 1
-            `,
-		},
-	}
-
-	// Run tests.
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			// Clear the metrics after the test.
-			metrics.AdmissionRejectionsTotal.Reset()
-
-			// Call the function.
-			recordAdmissionRejection(tc.reason)
-
-			if err := testutil.GatherAndCompare(metrics.GetGather(), strings.NewReader(tc.wants), "kubelet_admission_rejections_total"); err != nil {
-				t.Error(err)
 			}
 		})
 	}
